requirements.txt
requests
pillow
beautifulsoup4
